// KitchenSinkAngular6/vendor/rappid.d.ts
export namespace g {

    export interface PlainPoint {

        x: number;
        y: number;
    }

    export interface PlainRect {

        x: number;
        y: number;
        width: number;
        height: number;
    }
// ...
}
// ...
export namespace dia {

    type Event = JQuery.TriggeredEvent;

    type Point = g.PlainPoint;

    type BBox = g.PlainRect;

    type Size = Pick<BBox, 'width' | 'height'>;

    type PaddingJSON = {
        left?: number;
        top?: number;
        right?: number;
        bottom?: number;
    };

    type Padding = number | PaddingJSON;
// ...
    export namespace Graph {

        interface Options {
            [key: string]: any;
        }

        interface ConnectionOptions extends Cell.EmbeddableOptions {
            inbound?: boolean;
            outbound?: boolean;
        }

        interface ExploreOptions extends ConnectionOptions {
            breadthFirst?: boolean;
        }
    }
// ...
    export namespace Cell {

        interface GenericAttributes<T> {
            attrs?: T;
            z?: number;
            [key: string]: any;
        }

        interface Selectors {
            [selector: string]: attributes.SVGAttributes | undefined;
        }

        interface Attributes extends GenericAttributes<Selectors> {
            [key: string]: any;
        }

        interface Constructor<T extends Backbone.Model> {
            new (opt?: { id: string }): T
        }

        interface Options {
            [key: string]: any;
        }

        interface EmbeddableOptions extends Options {
            deep?: boolean;
        }

        interface DisconnectableOptions extends Options {
            disconnectLinks?: boolean;
        }

        interface TransitionOptions extends Options {
            delay?: number;
            duration?: number;
            timingFunction?: util.timing.TimingFunction;
            valueFunction?: util.interpolate.InterpolateFunction<any>;
        }
    }
// ...
    export namespace CellView {

        interface Options<T extends Cell> extends mvc.ViewOptions<T> {
            id?: string
        }

        interface InteractivityOptions extends ElementView.InteractivityOptions, LinkView.InteractivityOptions {

        }

        type FlagLabel = string | string[];
        type PresentationAttributes = { [key: string]: FlagLabel };
    }
// ...
    export namespace ElementView {

        interface InteractivityOptions {
            elementMove?: boolean;
            addLinkFromMagnet?: boolean;
        }
    }
// ...
    export namespace LinkView {

        interface InteractivityOptions {
            vertexAdd?: boolean,
            vertexMove?: boolean,
            vertexRemove?: boolean;
            arrowheadMove?: boolean;
            labelMove?: boolean;
            linkMove?: boolean;
            useLinkTools?: boolean;
        }

        interface GetConnectionPoint {
            (
                linkView: LinkView,
                view: ElementView,
                magnet: SVGElement,
                reference: Point,
                end: LinkEnd
            ): Point;
        }

        interface LabelOptions extends Cell.Options {
            absoluteDistance?: boolean;
            reverseDistance?: boolean;
            absoluteOffset?: boolean;
            keepGradient?: boolean;
            ensureLegibility?: boolean;
        }

        interface VertexOptions extends Cell.Options {

        }

        interface Options extends mvc.ViewOptions<Link> {
            shortLinkLength?: number,
            doubleLinkTools?: boolean,
            longLinkLength?: number,
            linkToolsOffset?: number,
            doubleLinkToolsOffset?: number,
            sampleInterval?: number
        }
    }
// ...
}
// ...
export namespace mvc {

    interface ViewOptions<T extends Backbone.Model> extends Backbone.ViewOptions<T> {
        theme?: string;
    }

    interface viewEventData {
        [key: string]: any;
    }

    class View<T extends Backbone.Model> extends Backbone.View<T> {

        constructor(opt?: ViewOptions<T>);

        UPDATE_PRIORITY: number;

        vel: Vectorizer | null;

        options: ViewOptions<T>;

        theme: string;

        themeClassNamePrefix: string

        defaultTheme: string;

        requireSetThemeOverride: boolean;

        documentEvents?: Backbone.EventsHash;

        children?: dia.MarkupJSON;

        childNodes?: { [key: string]: Element } | null;

        setTheme(theme: string, opt?: { override?: boolean }): this;

        getEventNamespace(): string;

        delegateDocumentEvents(events?: Backbone.EventsHash, data?: viewEventData): this;

        undelegateDocumentEvents(): this;

        delegateElementEvents(element: Element, events?: Backbone.EventsHash, data?: viewEventData): this;

        undelegateElementEvents(element: Element): this;

        eventData(evt: dia.Event): viewEventData;
        eventData(evt: dia.Event, data: viewEventData): this;

        stopPropagation(evt: dia.Event): this;
        isPropagationStopped(evt: dia.Event): boolean;

        renderChildren(children?: dia.MarkupJSON): this;

        findAttribute(attributeName: string, node: Element): string | null;

        confirmUpdate(flag: number, opt: { [key: string]: any }): number;

        unmount(): void;

        protected init(): void;

        protected onRender(): void;

        protected onSetTheme(oldTheme: string, newTheme: string): void;

        protected onRemove(): void;
    }
}
// ...
export namespace connectionPoints {

    interface DefaultConnectionPointArguments {
        offset?: number;
    }

    interface StrokeConnectionPointArguments extends DefaultConnectionPointArguments {
        stroke?: boolean;
    }

    interface BoundaryConnectionPointArguments extends StrokeConnectionPointArguments {
        selector?: Array<string | number> | string;
        precision?: number;
        extrapolate?: boolean;
        sticky?: boolean;
        insideout?: boolean;
    }

    interface ConnectionPointArgumentsMap {
        'anchor': DefaultConnectionPointArguments,
        'bbox': StrokeConnectionPointArguments,
        'rectangle': StrokeConnectionPointArguments,
        'boundary': BoundaryConnectionPointArguments,
        [key: string]: { [key: string]: any };
    }

    type ConnectionPointType = keyof ConnectionPointArgumentsMap;

    type GenericConnectionPointArguments<K extends ConnectionPointType> = ConnectionPointArgumentsMap[K];

    interface GenericConnectionPoint<K extends ConnectionPointType> {
        (
            endPathSegmentLine: g.Line,
            endView: dia.CellView,
            endMagnet: SVGElement,
            opt: ConnectionPointArgumentsMap[K],
            //endType: string,
            //linkView: dia.LinkView
        ): g.Point;
    }

    interface GenericConnectionPointJSON<K extends ConnectionPointType> {
        name: K;
        args?: ConnectionPointArgumentsMap[K];
    }

    type ConnectionPointArguments = GenericConnectionPointArguments<ConnectionPointType>;

    type ConnectionPoint = GenericConnectionPoint<ConnectionPointType>;

    type ConnectionPointJSON = GenericConnectionPointJSON<ConnectionPointType>;

    export var anchor: GenericConnectionPoint<'anchor'>;
    export var bbox: GenericConnectionPoint<'bbox'>;
    export var rectangle: GenericConnectionPoint<'rectangle'>;
    export var boundary: GenericConnectionPoint<'boundary'>;
}
// ...
export namespace dia {

    export namespace Paper {

        interface PrintExportOptions {
            size?: string;
            padding?: dia.Padding;
        }

        type BeforeSerialize = (doc: SVGSVGElement, paper: dia.Paper) => void | SVGElement;

        interface RasterExportOptions {
            type?: 'image/png' | 'image/jpeg' | string;
            height?: number;
            width?: number;
            size?: string;
            backgroundColor?: string;
            quality?: number;
            padding?: dia.Padding;
            area?: dia.BBox;
            useComputedStyles?: boolean;
            stylesheet?: string;
            canvg?: any;
            beforeSerialize?: BeforeSerialize;
        }

        interface SVGExportOptions {
            preserveDimensions?: boolean;
            area?: dia.BBox;
            convertImagesToDataUris?: boolean;
            useComputedStyles?: boolean;
            stylesheet?: string;
            beforeSerialize?: BeforeSerialize;
        }
    }

    export interface Paper {

        print(opt?: Paper.PrintExportOptions): void;

        toDataURL(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toJPEG(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toPNG(callback: (dataURL: string) => void, opt?: Paper.RasterExportOptions): void;

        toSVG(callback: (svg: string) => void, opt?: Paper.SVGExportOptions): void;

        openAsSVG(opt?: Paper.SVGExportOptions): void;

        openAsPNG(opt?: Paper.RasterExportOptions): void;
    }

    namespace CommandManager {

        export interface Options {
            graph: dia.Graph;
            cmdBeforeAdd?: (eventName: string, ...eventArgs: any[]) => boolean;
            cmdNameRegex?: any; /* a regular expression */
            applyOptionsList?: string[];
            revertOptionsList?: string[];
        }
    }

    class CommandManager extends Backbone.Model {

        constructor(opt: CommandManager.Options);

        undo(opt?: { [key: string]: any }): void;

        redo(opt?: { [key: string]: any }): void;

        cancel(opt: { [key: string]: any }): void;

        reset(): void;

        hasUndo(): boolean;

        hasRedo(): boolean;

        listen(): void;

        initBatchCommand(): void;

        storeBatchCommand(): void;
    }

    namespace Validator {

        export interface Options {
            commandManager: dia.CommandManager;
            cancelInvalid?: boolean;
        }
    }

    class Validator extends Backbone.Model {

        constructor(opt: Validator.Options);

        validate(actions: string, ...callback: Array<(err: Error, command: any, next: any) => boolean>): Validator;
    }
}